refactoring in progress / done for basic features
